requirements.txt
fastapi==0.104.1 
uvicorn[standard]==0.24.0 
python-multipart==0.0.6 
aiofiles==23.2.1 
python-docx==1.1.0 
PyPDF2==3.0.1 
requests==2.31.0 
python-jose[cryptography]==3.3.0 
passlib[bcrypt]==1.7.4 
Pillow==10.0.1 
